+++ Examples/Examples-iOS/IGListKitExamples/SectionControllers/SearchSectionController.swift
@@ -32,31 +32,34 @@ final class SearchSectionController: ListSectionController, UISearchBarDelegate,
     }
 
     override func cellForItem(at index: Int) -> UICollectionViewCell {
-        let cell = collectionContext!.dequeueReusableCell(of: SearchCell.self, for: self, at: index) as! SearchCell
+        guard let cell = collectionContext?.dequeueReusableCell(of: SearchCell.self, for: self, at: index) as? SearchCell else {
+            fatalError()
+        }
         cell.searchBar.delegate = self
         return cell
     }
 
-    //MARK: UISearchBarDelegate
+    // MARK: UISearchBarDelegate
 
     func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
         delegate?.searchSectionController(self, didChangeText: searchText)
     }
 
     func searchBarTextDidEndEditing(_ searchBar: UISearchBar) {
-        delegate?.searchSectionController(self, didChangeText: "")
+        delegate?.searchSectionController(self, didChangeText: searchBar.text!)
     }
 
-    //MARK: ListScrollDelegate
+    // MARK: ListScrollDelegate
 
     func listAdapter(_ listAdapter: ListAdapter, didScroll sectionController: ListSectionController) {
         if let searchBar = (collectionContext?.cellForItem(at: 0, sectionController: self) as? SearchCell)?.searchBar {
-            searchBar.text = ""
             searchBar.resignFirstResponder()
         }
     }
 
     func listAdapter(_ listAdapter: ListAdapter, willBeginDragging sectionController: ListSectionController) {}
-    func listAdapter(_ listAdapter: ListAdapter, didEndDragging sectionController: ListSectionController, willDecelerate decelerate: Bool) {}
+    func listAdapter(_ listAdapter: ListAdapter,
+                     didEndDragging sectionController: ListSectionController,
+                     willDecelerate decelerate: Bool) {}
 
 }