+++ Source/IGListAdapter.m
@@ -18,6 +18,8 @@
 
 @implementation IGListAdapter {
     NSMapTable<UICollectionReusableView *, IGListSectionController *> *_viewSectionControllerMap;
+    // An array of blocks to execute once batch updates are finished
+    NSMutableArray<void (^)()> *_queuedCompletionBlocks;
 }
 
 - (void)dealloc {
@@ -46,7 +48,7 @@
         NSMapTable *table = [[NSMapTable alloc] initWithKeyPointerFunctions:keyFunctions valuePointerFunctions:valueFunctions capacity:0];
         _sectionMap = [[IGListSectionMap alloc] initWithMapTable:table];
 
-        _displayHandler = [[IGListDisplayHandler alloc] init];
+        _displayHandler = [IGListDisplayHandler new];
         _workingRangeHandler = [[IGListWorkingRangeHandler alloc] initWithWorkingRangeSize:workingRangeSize];
 
         _viewSectionControllerMap = [NSMapTable mapTableWithKeyOptions:NSMapTableObjectPointerPersonality | NSMapTableStrongMemory
@@ -96,6 +98,11 @@
 
         _collectionView = collectionView;
         _collectionView.dataSource = self;
+
+        if ([_collectionView respondsToSelector:@selector(setPrefetchingEnabled:)]) {
+            _collectionView.prefetchingEnabled = NO;
+        }
+
         [_collectionView.collectionViewLayout invalidateLayout];
 
         [self updateCollectionViewDelegate];
@@ -172,20 +179,41 @@
     }
 
     UICollectionView *collectionView = self.collectionView;
-    const NSInteger numberOfItems = [collectionView numberOfItemsInSection:section];
-    if (numberOfItems == 0) {
-        return;
-    }
+    UICollectionViewLayout *layout = self.collectionView.collectionViewLayout;
 
     // force layout before continuing
     // this method is typcially called before pushing a view controller
     // thus, before the layout process has actually happened
+    [collectionView setNeedsLayout];
     [collectionView layoutIfNeeded];
 
+    NSIndexPath *indexPathFirstElement = [NSIndexPath indexPathForItem:0 inSection:section];
+    
     // collect the layout attributes for the cell and supplementary views for the first index
     // this will break if there are supplementary views beyond item 0
-    NSIndexPath *indexPath = [NSIndexPath indexPathForItem:0 inSection:section];
-    NSArray *attributes = [self layoutAttributesForIndexPath:indexPath supplementaryKinds:supplementaryKinds];
+    NSMutableArray<UICollectionViewLayoutAttributes *> *attributes = nil;
+    
+    const NSInteger numberOfItems = [collectionView numberOfItemsInSection:section];
+    if (numberOfItems > 0) {
+        attributes = [self layoutAttributesForIndexPath:indexPathFirstElement supplementaryKinds:supplementaryKinds].mutableCopy;
+
+        if (numberOfItems > 1) {
+            NSIndexPath *indexPathLastElement = [NSIndexPath indexPathForItem:(numberOfItems - 1) inSection:section];
+            UICollectionViewLayoutAttributes *lastElementattributes = [self layoutAttributesForIndexPath:indexPathLastElement supplementaryKinds:supplementaryKinds].firstObject;
+            if (lastElementattributes != nil) {
+                [attributes addObject:lastElementattributes];
+            }
+        }
+    } else {
+        NSMutableArray *supplementaryAttributes = [NSMutableArray new];
+        for (NSString* supplementaryKind in supplementaryKinds) {
+            UICollectionViewLayoutAttributes *supplementaryAttribute = [layout layoutAttributesForSupplementaryViewOfKind:supplementaryKind atIndexPath:indexPathFirstElement];
+            if (supplementaryAttribute != nil) {
+                [supplementaryAttributes addObject: supplementaryAttribute];
+            }
+        }
+        attributes = supplementaryAttributes;
+    }
 
     CGFloat offsetMin = 0.0;
     CGFloat offsetMax = 0.0;
@@ -220,7 +248,7 @@
     const UIEdgeInsets contentInset = collectionView.contentInset;
     CGPoint contentOffset = collectionView.contentOffset;
     switch (scrollDirection) {
-        case UICollectionViewScrollDirectionHorizontal:
+        case UICollectionViewScrollDirectionHorizontal: {
             switch (scrollPosition) {
                 case UICollectionViewScrollPositionRight:
                     contentOffset.x = offsetMax - collectionViewWidth - contentInset.left;
@@ -238,8 +266,13 @@
                     contentOffset.x = offsetMin - contentInset.left;
                     break;
             }
+            const CGFloat maxOffsetX = collectionView.contentSize.width - collectionView.frame.size.width + collectionView.contentInset.right;
+            const CGFloat minOffsetX = -collectionView.contentInset.left;
+            contentOffset.x = MIN(contentOffset.x, maxOffsetX);
+            contentOffset.x = MAX(contentOffset.x, minOffsetX);
             break;
-        case UICollectionViewScrollDirectionVertical:
+        }
+        case UICollectionViewScrollDirectionVertical: {
             switch (scrollPosition) {
                 case UICollectionViewScrollPositionBottom:
                     contentOffset.y = offsetMax - collectionViewHeight - contentInset.top;
@@ -257,12 +290,16 @@
                     contentOffset.y = offsetMin - contentInset.top;
                     break;
             }
+            const CGFloat maxOffsetY = collectionView.contentSize.height - collectionView.frame.size.height + collectionView.contentInset.bottom;
+            const CGFloat minOffsetY = -collectionView.contentInset.top;
+            contentOffset.y = MIN(contentOffset.y, maxOffsetY);
+            contentOffset.y = MAX(contentOffset.y, minOffsetY);
             break;
+        }
     }
 
     [collectionView setContentOffset:contentOffset animated:animated];
 }
-
 
 #pragma mark - Editing
 
@@ -272,6 +309,7 @@
     id<IGListAdapterDataSource> dataSource = self.dataSource;
     UICollectionView *collectionView = self.collectionView;
     if (dataSource == nil || collectionView == nil) {
+        IGLKLog(@"Warning: Your call to %s is ignored as dataSource or collectionView haven't been set.", __PRETTY_FUNCTION__);        
         if (completion) {
             completion(NO);
         }
@@ -280,6 +318,8 @@
 
     NSArray *fromObjects = self.sectionMap.objects;
     NSArray *newObjects = [dataSource objectsForListAdapter:self];
+
+    [self enterBatchUpdates];
 
     __weak __typeof__(self) weakSelf = self;
     [self.updater performUpdateWithCollectionView:collectionView
@@ -299,6 +339,8 @@
                                 if (completion) {
                                     completion(finished);
                                 }
+
+                                [weakSelf exitBatchUpdates];
                             }];
 }
 
@@ -308,6 +350,7 @@
     id<IGListAdapterDataSource> dataSource = self.dataSource;
     UICollectionView *collectionView = self.collectionView;
     if (dataSource == nil || collectionView == nil) {
+        IGLKLog(@"Warning: Your call to %s is ignored as dataSource or collectionView haven't been set.", __PRETTY_FUNCTION__);
         if (completion) {
             completion(NO);
         }
@@ -328,7 +371,7 @@
     IGAssertMainThread();
     IGParameterAssert(objects);
 
-    NSMutableIndexSet *sections = [[NSMutableIndexSet alloc] init];
+    NSMutableIndexSet *sections = [NSMutableIndexSet new];
 
     // use the item map based on whether or not we're in an update block
     IGListSectionMap *map = [self sectionMapUsingPreviousIfInUpdateBlock:YES];
@@ -412,11 +455,12 @@
 
 - (NSArray<IGListSectionController *> *)visibleSectionControllers {
     IGAssertMainThread();
-    NSArray<UICollectionViewCell *> *visibleCells = [self.collectionView visibleCells];
     NSMutableSet *visibleSectionControllers = [NSMutableSet new];
-    for (UICollectionViewCell *cell in visibleCells) {
-        IGListSectionController *sectionController = [self sectionControllerForView:cell];
-        IGAssert(sectionController != nil, @"Section controller nil for cell %@", cell);
+    NSArray<UICollectionViewLayoutAttributes *> *attributes =
+    [self.collectionView.collectionViewLayout layoutAttributesForElementsInRect:self.collectionView.bounds];
+    for (UICollectionViewLayoutAttributes* attribute in attributes) {
+        IGListSectionController *sectionController = [self sectionControllerForSection:attribute.indexPath.section];
+        IGAssert(sectionController != nil, @"Section controller nil for cell in section %zd", attribute.indexPath.section);
         if (sectionController) {
             [visibleSectionControllers addObject:sectionController];
         }
@@ -492,8 +536,11 @@
     IGParameterAssert(dataSource != nil);
 
 #if DEBUG
+    NSCountedSet *identifiersSet = [NSCountedSet new];
     for (id object in objects) {
+        [identifiersSet addObject:[object diffIdentifier]];
         IGAssert([object isEqualToDiffableObject:object], @"Object instance %@ not equal to itself. This will break infra map tables.", object);
+        IGAssert([identifiersSet countForObject:[object diffIdentifier]] <= 1, @"Diff identifier %@ for object %@ occurs more than once. Identifiers must be unique!", [object diffIdentifier], object);
     }
 #endif
 
@@ -600,7 +647,7 @@
 - (NSArray<NSIndexPath *> *)indexPathsFromSectionController:(IGListSectionController *)sectionController
                                                     indexes:(NSIndexSet *)indexes
                                  usePreviousIfInUpdateBlock:(BOOL)usePreviousIfInUpdateBlock {
-    NSMutableArray<NSIndexPath *> *indexPaths = [[NSMutableArray alloc] init];
+    NSMutableArray<NSIndexPath *> *indexPaths = [NSMutableArray new];
 
     IGListSectionMap *map = [self sectionMapUsingPreviousIfInUpdateBlock:usePreviousIfInUpdateBlock];
     const NSInteger section = [map sectionForSectionController:sectionController];
@@ -627,7 +674,7 @@
 - (NSArray<UICollectionViewLayoutAttributes *> *)layoutAttributesForIndexPath:(NSIndexPath *)indexPath
                                                            supplementaryKinds:(NSArray<NSString *> *)supplementaryKinds {
     UICollectionViewLayout *layout = self.collectionView.collectionViewLayout;
-    NSMutableArray<UICollectionViewLayoutAttributes *> *attributes = [[NSMutableArray alloc] init];
+    NSMutableArray<UICollectionViewLayoutAttributes *> *attributes = [NSMutableArray new];
 
     UICollectionViewLayoutAttributes *cellAttributes = [layout layoutAttributesForItemAtIndexPath:indexPath];
     if (cellAttributes) {
@@ -659,6 +706,26 @@
 - (void)removeMapForView:(UICollectionReusableView *)view {
     IGAssertMainThread();
     [_viewSectionControllerMap removeObjectForKey:view];
+}
+
+- (void)deferBlockBetweenBatchUpdates:(void (^)())block {
+    IGAssertMainThread();
+    if (_queuedCompletionBlocks == nil) {
+        block();
+    } else {
+        [_queuedCompletionBlocks addObject:block];
+    }
+}
+
+- (void)enterBatchUpdates {
+    _queuedCompletionBlocks = [NSMutableArray new];
+}
+
+- (void)exitBatchUpdates {
+    for (void (^block)() in _queuedCompletionBlocks) {
+        block();
+    }
+    _queuedCompletionBlocks = nil;
 }
 
 #pragma mark - UIScrollViewDelegate
@@ -707,7 +774,10 @@
     }
     NSArray<IGListSectionController *> *visibleSectionControllers = [self visibleSectionControllers];
     for (IGListSectionController *sectionController in visibleSectionControllers) {
-        [[sectionController scrollDelegate] listAdapter:self didEndDeceleratingSectionController:sectionController];
+        id<IGListScrollDelegate> scrollDelegate = [sectionController scrollDelegate];
+        if ([scrollDelegate respondsToSelector:@selector(listAdapter:didEndDeceleratingSectionController:)]) {
+            [scrollDelegate listAdapter:self didEndDeceleratingSectionController:sectionController];
+        }
     }
 }
 
@@ -737,7 +807,8 @@
     IGParameterAssert(cell != nil);
     IGParameterAssert(sectionController != nil);
     NSIndexPath *indexPath = [self.collectionView indexPathForCell:cell];
-    IGAssert(indexPath.section == [self sectionForSectionController:sectionController],
+    IGAssert(indexPath == nil
+             || indexPath.section == [self sectionForSectionController:sectionController],
              @"Requesting a cell from another section controller is not allowed.");
     return indexPath != nil ? indexPath.item : NSNotFound;
 }
@@ -799,6 +870,16 @@
     IGParameterAssert(sectionController != nil);
     NSIndexPath *indexPath = [self indexPathForSectionController:sectionController index:index usePreviousIfInUpdateBlock:NO];
     [self.collectionView deselectItemAtIndexPath:indexPath animated:animated];
+}
+
+- (void)selectItemAtIndex:(NSInteger)index
+        sectionController:(IGListSectionController *)sectionController
+                 animated:(BOOL)animated
+            scrollPosition:(UICollectionViewScrollPosition)scrollPosition {
+    IGAssertMainThread();
+    IGParameterAssert(sectionController != nil);
+    NSIndexPath *indexPath = [self indexPathForSectionController:sectionController index:index usePreviousIfInUpdateBlock:NO];
+    [self.collectionView selectItemAtIndexPath:indexPath animated:animated scrollPosition:scrollPosition];
 }
 
 - (__kindof UICollectionViewCell *)dequeueReusableCellOfClass:(Class)cellClass
@@ -909,6 +990,8 @@
     IGParameterAssert(updates != nil);
     UICollectionView *collectionView = self.collectionView;
     IGAssert(collectionView != nil, @"Performing batch updates without a collection view.");
+
+    [self enterBatchUpdates];
     
     __weak __typeof__(self) weakSelf = self;
     [self.updater performUpdateWithCollectionView:collectionView animated:animated itemUpdates:^{
@@ -921,6 +1004,8 @@
         if (completion) {
             completion(finished);
         }
+
+        [weakSelf exitBatchUpdates];
     }];
 }
 
@@ -949,11 +1034,14 @@
     UICollectionViewLayoutInvalidationContext *context = [[[layout.class invalidationContextClass] alloc] init];
     [context invalidateItemsAtIndexPaths:indexPaths];
     
-    void (^block)() = ^{
-        [layout invalidateLayoutWithContext:context];
-    };
-    
-    [_collectionView performBatchUpdates:block completion:completion];
+    __weak __typeof__(_collectionView) weakCollectionView = _collectionView;
+
+    // do not call -[UICollectionView performBatchUpdates:completion:] while already updating. defer it until completed.
+    [self deferBlockBetweenBatchUpdates:^{
+        [weakCollectionView performBatchUpdates:^{
+            [layout invalidateLayoutWithContext:context];
+        } completion:completion];
+    }];
 }
 
 #pragma mark - IGListBatchContext