+++ Tests/Objects/IGListTestSection.m
@@ -46,4 +46,16 @@
     self.wasSelected = YES;
 }
 
+- (void)didDeselectItemAtIndex:(NSInteger)index {
+    self.wasDeselected = YES;
+}
+
+- (void)didHighlightItemAtIndex:(NSInteger)index {
+    self.wasHighlighted = YES;
+}
+
+- (void)didUnhighlightItemAtIndex:(NSInteger)index {
+    self.wasUnhighlighted = YES;
+}
+
 @end