+++ Tests/Objects/IGTestDiffingSectionController.m
@@ -55,4 +55,16 @@
     self.selectedViewModel = viewModel;
 }
 
+- (void)sectionController:(IGListBindingSectionController *)sectionController didDeselectItemAtIndex:(NSInteger)index viewModel:(id)viewModel {
+    self.deselectedViewModel = viewModel;
+}
+
+- (void)sectionController:(IGListBindingSectionController *)sectionController didHighlightItemAtIndex:(NSInteger)index viewModel:(id)viewModel {
+    self.highlightedViewModel = viewModel;
+}
+
+- (void)sectionController:(IGListBindingSectionController *)sectionController didUnhighlightItemAtIndex:(NSInteger)index viewModel:(id)viewModel {
+    self.unhighlightedViewModel = viewModel;
+}
+
 @end